fix a few warnings in 32-bit builds

+++ tutorials/interceptlayer/main-part2solution.cpp
@@ -130,7 +130,7 @@ static void go()
     commandQueue.finish();
 
     auto start = std::chrono::system_clock::now();
-    for( int i = 0; i < iterations; i++ )
+    for( size_t i = 0; i < iterations; i++ )
     {
         commandQueue.enqueueNDRangeKernel(
             kernel,
